backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
bcrypt==4.1.2 
python-multipart==0.0.6
pydantic[email]==2.7.0
python-dotenv==1.0.0
requests==2.31.0